Auto-deploy via Travis
Summary:
First attempt at automation of wheel & website deployment. This is a draft, but we do need to push it, because Travis does not allow to trigger a custom build on a tagged commit (it's possible, but then the tagged commit is used as branch and not actually treated as tagged –– https://github.com/travis-ci/beta-features/issues/27#issuecomment-463881983 is reporting this issue).

Committing this will not result in deployment as deployment is set up to 'Test PyPI' and `gh-pages` pushing is disabled in the website-publishing script.

Reviewed By: kkashin

Differential Revision: D17330229

fbshipit-source-id: 470b57c88c3871a3dd476aed6f479fe926e1de14