Fixing small storage bug
Summary: We still have inconsistencies in our code with numpy vs. regular Python types. Ideally we'd be using regular Python types everywhere we can, but it seems that's not exactly happening and it's pretty tricky to track down every case. Instead of relying on that when we check type equality in our storage code, let's just cast to Python types.

Reviewed By: lena-kashtelyan

Differential Revision: D17391040

fbshipit-source-id: ed13ca85ef68de3695a6ee75139e984fd1541883