Handle OOD points in base modelbridge
Summary:
OOD Points should be handled in one place and one place only.
This diff makes that place the `__init__` method of base modelbridge, and removes the old OOD transform.

Before this lands, I will make sure to drop all saved instances of this transform in the new AX DB.

Note: I've done some bad things (read: pragma no cover) in some edge cases of the modelbridge tests.
I would like to cover these, but it's very difficult in the current test structure. I will need to significantly re-write tests to get coverage here, and I think that will be a project in and of itself.

Reviewed By: bletham

Differential Revision: D17412826

fbshipit-source-id: 1dcafa1380b939bf5c81cd257ed3ad3a15cbc2c1