Display timestamp when using cached data
Summary:
In T54069829 (D17554955), we added a message to the top of the report when we use cached data.

We are adding to this message to display a **timestamp** so users will know exactly how old this cached data is.

To get the storage time of the cached data, I modified `lookup_data_for_trial` (https://our.intern.facebook.com/intern/diffusion/FBS/browse/master/fbcode/ax/ax/core/experiment.py?lines=419) to return a tuple of the `Data` object and the timestamp in milliseconds. As a result, I had to modify some other places in the code that called `lookup_data_for_trial` and ignore the timestamp returned.

Screenshot of info message with date and time:

{F213403738}

Reviewed By: sdsingh

Differential Revision: D17668530

fbshipit-source-id: 46db20fd9180c7877c0f6e9e221500226d5d6f53