Allow for finding last trial with parameterization(s) for Maximus
Summary:
Methods for 1) finding the last trial in experiment with given parameterization(s) and 2) verifying that a trial index corresponds to a parameterization of an arm correctly.

Both are needed for Maximus; making 1) private since hopefully OSS users will not need that functionality often and 2) public, because that functionality should generally be helpful in a distributed environment where the external system running the trials can get out-of-sync.

Reviewed By: ldworkin

Differential Revision: D17750028

fbshipit-source-id: 5b5cf211d37de12c89cf9376c852205dfdd14f52