Base context class
Summary:
Work in progress, looking for comments.

Create a basic context class. This is a very cut down version of what the old CC/AE context class was, but should support most of the same things.

There are two things this won't support natively:
1. Regex based buckets/parameters.
2. Exclude based buckets/parameters: NOT IN, !=, etc.

From the Ax side, it seems possible that we might be able to treat these the same way as catch all, and just create dummy parameters.

For reference: https://fburl.com/cc_context

On the other hand, I can see an argument for separating buckets out from parameters--curious to hear others thoughts!

Reviewed By: lena-kashtelyan

Differential Revision: D16060688

fbshipit-source-id: c6eeae62648afd5c443feca433980904e7b870a0