Adding "additional_experiment" to MetricIntent enum
Summary:
We have this type in PHP, thanks to Jake Podell, but not in Python, and this was causing some SQL alchemy issues (see test plan).

It's not *great* that this goes into the OSS code, since we don't actually have support for additional objectives in  Python, but I personally didn't think it was a big enough deal to warrant a FB internal version of this enum. Feel free to push back on me if you think we should.

Reviewed By: lena-kashtelyan

Differential Revision: D18007237

fbshipit-source-id: 576661860e16a620767fc86647324aff48e71bf5