LogY transform
Summary:
Adds a `LogY` transform to the Ax transforms for log-transforming observations.
This is useful if for log-normally distributed outcomes, and if there is multiplicative noise.
See https://fb.workplace.com/groups/aeeng/permalink/531165437642157/ for more details.

This implements two modes:

Delta method:
Turns out there is a bijective transform from (R^n, S_+^n) to (R^n, S_+^n) between the mean and covariance matrix of a log-Normal and that of it's underlying Normal parameterization (bijective between mean and covariance, not the full distribution of course). We use that transformation (for now in a degenerate case because we only look at variances, but this works for full covariances as well).

"CI matching":
- in the transform direction, if there is observation nose, we compute the
  symmetric CI under the normality assumption, then transform the end of this
  CI to the `log` space and compute its width. We then back out the variance in
  the `log` space that, when interpreted again as a normal symmetric CI of the
  same level, has the same width.
- in the untransform direction, we do the same in reverse, i.e. just use
  `exp` instead of the `log1` transform.

Turns out that these do pretty much the same thing for the example in the test plan. If there are full covariances involved they should behave differently.

**Notes:**

1. A more reasonable way of handling this would be to support non-symmetric CIs in Ax that are described by a `center`, and a `(lower, upper)` tuple. See also https://fb.workplace.com/groups/aeeng/permalink/531165437642157/ for some discussion on this.
2. For log-transformed outcomes, we probably also want to allow the plots to be on a log scale (see the runtime_averaged metric in the CV plot  in the test nb, which is quite squashed together).

Reviewed By: bletham

Differential Revision: D17983495

fbshipit-source-id: 923361431a7bbf0cb2d1c059c3c5999ab4bfcad8