Adding `model_state_after_gen` to generator run
Summary:
In order to not use `update` but only use `fit` in generation strategy, we need to keep track of stateful models' (e.g. Sobol) states. An easy way of doing this is by saving them on the generator run.

For instance, Sobol's state consists of 1 value currently, its initial position (might contain more things, like seed, in the future). We will save that value on the generator run, so that when restoring the model from generator run, we can restore its state after its last generation, and resume the optimization.

Reviewed By: bletham

Differential Revision: D18406915

fbshipit-source-id: b2064476b2c5554566a6803860e33e77b1844711