Get EB-modeled data from analysis
Summary:
The next phase of my project involves porting over the **Generate Candidates** cell of the reports to intern. The sufficient statistics we need to send to the front-end are to replace the HTML blob:

1) EB-modeled data
2) GP-modeled data
3) Model diagnostics

For more context, please refer to this Quip doc: https://fb.quip.com/dstaADntP65b

This diff:
- Includes data structures for EB-modeled data (as Thrift structs)
- Adds a function `gen_eb_predictions()` to get EB-modeled data and ReportErrorMessages from analysis report inputs and data

My next stacked diff add EB-modeled data as a field in `AnalysisStatistics`, and ensure EB-modeled data can be successfully saved in the database. I wanted to split the diffs up for ease of review :)

Reviewed By: ldworkin

Differential Revision: D18399758

fbshipit-source-id: 04cf4ba0ffdcc2a42ded6627494c8220b2237d27