Basic functions
Summary:
To make the benchmarking code more modular (basically reduces it to 4 functions), here it's all changed to be functional (away with `BenchmarkRunner` and `BenchmarkSuite` classes).

This diff also temporarily removed the benchmarking tutorial, as it will need to be rewritten, which I'll do a bit later on.

Next steps:
- FBL workflow: D18009570, test run: f146606026
- add unit tests
- propagate aggregated error messages to the report
- add more evaluation procedures
- batched benchmarking
- simulate async evaluation
- update the tutorial on benchmarking

Reviewed By: bletham

Differential Revision: D17929709

fbshipit-source-id: 0e73059e29d93a0585b116cc2e08f45b969513c1