Use `modelbridge.fit` instead or `update` in generation strategy
Summary:
Problem: when saved, Ax client forgets it's `_updated_trials` field, on which it relies to get new data (that the generation strategy has not yet seen). When reinstantiated, `_updated_trials` is empty, because of which no data ends up being passed into the generation strategy. This is actually very bad and I'm really glad this got surfaced in Kevin's testing of Maximus with HHVM

Solution: luckily, we recently moved away from using `model_bridge.update` (which required the data passed to the generation strategy, to be data the GS hasn't seen yet) method anywhere. Therefore, we can simply throw away `_updated_trials`, move from using `update` to just using `fit` in the GS, and not need to update any of the storage to start including `_updated_trials`.

Reviewed By: bletham

Differential Revision: D18333985

fbshipit-source-id: 998eb31dbde90b9ca92bf387ddf70a0809655e89