Properly handle target values in transforms
Summary:
Currently, target values of fidelity parameters (and for that matter, fidelity parameters more generally) are not properly handled in the Ax transforms.
This diff makes sure that any transform that modifies the search space also properly handles target values of fidelity parameters.
It also adds some additional errors to be raised when we attempt things with fidelity parameters that we should not.

Differential Revision: D18648222

fbshipit-source-id: 387e0c029e2bdf1fd125e92d47338a5b195200dd