Properly handle target values in transforms (#210)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/210

Currently, target values of fidelity parameters (and for that matter, fidelity parameters more generally) are not properly handled in the Ax transforms.
This diff makes sure that any transform that modifies the search space also properly handles target values of fidelity parameters.
It also adds some additional errors to be raised when we attempt things with fidelity parameters that we should not.

Reviewed By: bletham

Differential Revision: D18648222

fbshipit-source-id: a6e9426d05dc74d8cd65739917094ce3ec727f1c