Support mixing absolute and relativized data in scatter plots.
Summary:
Adaptive GraphQL wants to optimize on absolute constraints on gcpu.
This request has also come up in the past, and is why we support absolute constraints.

Unfortunately, our plotting does not support this very well. To do so, we need the ability to specify the `rel` flag separately for each metric in a plot.

This is not complete, but it does provide the necessary changes for all scatter plots to implement this.
It implements this for one particularly useful scatter plot, `plot_objective_vs_constraints`

Reviewed By: 2timesjay

Differential Revision: D18627346

fbshipit-source-id: a9fafa4f0f590e667bf74ffd1a4a1273ecb8a685