Use unittest.mock instead of mock
Summary:
Do not use `import mock`. This uses a backported package of `unittest.mock`.
On recent python versions there is no reason to use `import mock`, ever.
In fact, it messes things up for OSS with a `ModuleNotFoundError`.

Not sure where this dep is being pulled in internally, but we should get rid of it to avoid this in the future.

Reviewed By: sdaulton

Differential Revision: D18814935

fbshipit-source-id: afab91c73921131841765519ed32ce3e9615e2a9