Make `overwrite_existing_experiment` in `create_experiment` fail if experiment set locally
Summary:
OSS user's PR description:
"Overwrite flag on `AxClient.create_experiment` should check for local experiments in addition  of DB experiments. This helps prevent accidental experiment deletion for use cases where the internal DB is not used as the persistent layer.

I added a test as I think it should behave like. Not sure if passes or not as I did not have it set up to run tests locally."

Lena's addition: I added the re-setting of generation strategy in cases

Reviewed By: sdsingh

Differential Revision: D18763389

fbshipit-source-id: 0d1f37bf7dbd2552359651b615777220795d42e7