Add ability to subset model along outcome dimensions (#226)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/226

Candidate generation currently by default uses a full model for all outcomes, even those not appearing in the OptimizationConfig. This is very wasteful in the current implementation, and leads to high memory consumption and long generation times (=this is likely what causes #202). See https://github.com/facebook/Ax/issues/205.

This diff provides a skeleton for subsetting the model to only those outcome dimensions that are needed for optimization during the generation step (i.e.those outcomes that have nonzero weight in `objective_weights` or `outcome_constraints`.

Right now this is just a skeleton, and will require implementing a `subset_output` method on the botorch models. The diff is written such that even without this being implemented on the model Ax falls back gracefully to the existing (wasteful) implementation.

Reviewed By: sdaulton

Differential Revision: D18666578

fbshipit-source-id: e7a8af17ac83381b9a7b88a46a5795d97b3a6cc7