Fix human in the loop tutorial
Summary:
D17577805 changes status_quo_weight to 'status_quo_weight_override', and adds 'optimize_for_power' (which is optional, but still required in the json?). We just conform the tutorial example to the new json definition.  I set optimize_for_power to false, if this is wrong we can change to true.

I intentionally didn't copy and paste the whole file, not sure how to get that diff to appear better. Is it part of the steps for diffing out tutorials?

Reviewed By: sdsingh

Differential Revision: D18937156

fbshipit-source-id: 9f178d25668db6aac51aada3ed351b0b38ae0db6