Remove logging when using callable in GS
Summary: This prints when doing a benchmark with a model not in the registry. Before it would only print once and it was fine. Now since GS was changed to instantiate the model in every iteration, it prints in every iteration and is a bit obnoxious.  See n187021 for an example with only 7 iterations and imagine what it looks like with a hundred...

Reviewed By: lena-kashtelyan

Differential Revision: D19187591

fbshipit-source-id: 74eeb26e012a0ee5b93adfba6b7ed447bd91cd0a