Make loading experiment compatible with removal of `DISPATCHED` status
Summary:
When loading an experiment from DB or JSON, some trials will come back with status `DISPATCHED`, which we will swap for `RUNNING` on decode to ensure backwards compatibility.

Without this change, trials with status `DISPATCHED` cannot be completed. When `DISPATCHED` was deprecated, it was essentially replaced by `RUNNING` everywhere it was used (in Service API). So this change is safe to make and will allow for completion of trials made before the `DISPATCHED` deprecation

Reviewed By: ldworkin

Differential Revision: D19292511

fbshipit-source-id: e51c93182a8564f2b0d72c99085de08cd7e3b6f6