Allow custom generation strategy in managed loop
Summary:
Right now custom generation strategies can only be used in the service API. It's a pretty simple change to allow for custom generation strategies in the managed loop API. This will make it much easier to use custom models on externally defined functions.

I'm particularly interested in this because it will allow for an easy interface to using the linear embedding technique for high-dimensional optimization.

Reviewed By: lena-kashtelyan

Differential Revision: D19333740

fbshipit-source-id: 298f73d0d353c8847fbf72d9da8f4918b262bacf