Add Percentile Transform
Summary:
Add a transform that maps Y values to their percentile in the distribution of other observations of the same metric.

For instance:

Observations [1,2,3,4], value: 2.5, result: 50.0.

See implementation here: https://github.com/scipy/scipy/blob/master/scipy/stats/stats.py#L1862

Please let me know if we want output in [0,1] or different handling of ties, or different handling of interpolation.

Also, if this is going to be put through an inverse gaussian cdf how should we handle 0.0 and 100.0 percentiles outcomes?

Reviewed By: sdsingh

Differential Revision: D19546496

fbshipit-source-id: a3ff9b257016746a91acfb96fe66674e85040f1c