Keep mapping of generation steps to trials
Summary:
This diff does quite a few things, but all with the intention of using Trial statuses to keep track of when the generation strategy should proceed to the next model (instead of parsing `Data` for new arms).

This way we:
1) get rid of hacky dataframe-parsing logic,
2) eliminate the need to Service / Loop API / Generation Strategy users to necessarily know what arms are and how they correlate with trials,
3) allow for a clean way of seeing all one's trials and the state of the generation strategy like so (source: n186960, is updated with the recent changes:

{F225321215}

Kostya and I thought that this was a good idea and made the code quite a bit cleaner; I'd love to hear what others think. The main tricky point (`_map_new_experiment_trials_to_steps`) has a long comment attached to it : )

NOTE: pls disregard the failing tests for now; looking for feedback on the logic itself, will fix tests later.

Reviewed By: ldworkin

Differential Revision: D19165422

fbshipit-source-id: 9cd320110074a474197f7869302c9ac309c62955