Set `batch_size` on `AxBatchClient` in Maximus
Summary:
Hardcode the value of 2 as batch size into Maximus to avoid needing to make another substantial change to Maximus in order to ship D19177898. Improving this is taskified: T61142544, and will be done before we start using Maximus for anything other than HHVM. keskici, please correct me if I'm wrong in understanding that that's our only use case and that the batch size there is always 2.

NOTE: none of the diffs in this stack can land separately; D19165422 fixes the imperfect logic left in D19165422, which allows the generation strategy to proceed forward too early (there's a place where it compares number of generated arms to number of required trials, which works fine for 1-arm trials but breaks for batched trials, causing the two Maximus test failures we see on this diff). If you didn't understand that lengthy explanation, no worries, it's on me to ensure that by the time the last diff in this stack (D19564405) is ready, all the tests will be passing.

Differential Revision: D19564405

fbshipit-source-id: ae3c98d7152b795551aa7e4609739de90d8b47d5