Store generation step information on generator run and trial
Summary:
In order to enable mapping generation steps to trials in generation strategy (done in D19165422), I'm  adding storage of generation step to generator run, from which its transferred to a trial when the generator run is added to one.

This is needed because at the time of `generation_strategy.gen`, a generator run (not a trial) is produced. Then, a trial is created with the generator run. So in order to keep a mapping of trials to generation step, a generation strategy (GS) needs to be able to update itself on each call to `gen`, to which an experiment is passed. GS parses the trials on that experiment for information about which generation steps produced each trial; it creates an up-to-date mapping of generation steps to trials they created. Then, GS uses that mapping to determine which step to generate from (and each step corresponds to the model).

Note: this code currently lacks storage support; before providing it, I wanted to ensure that my reviewers were on board with this diff and D19165422.

Reviewed By: ldworkin

Differential Revision: D19435226

fbshipit-source-id: 65b52c388f643c9413b9af338f5740a31da54908