Add reset_runners
Summary:
This is just an idea for an API that I find useful.
Frequently, when working in notebooks, a runner fails for some reason--either it is the wrong type of runner, or the runner implementation itself is wrong (during prototyping/dev). After that, it is difficult to fix/update batches with the faulty runner attached. This provides a utility to do so.

Reviewed By: 2timesjay

Differential Revision: D19837984

fbshipit-source-id: 4208526c567a397ee2748ff3b4ff2e94fc89d310