Genenerate experiment JSON file for human-in-the-loop tutorial
Summary: In D19435226, we added a generation_step_index property to `Trial` and `GeneratorRun`. These properties are missing, which is causing the error in the second cell of the human in the loop tutorial, where we load an experiment from a JSON file. This diff adds the missing property (with a value of `null`) to the sole trial of the JSON experiment.

Reviewed By: 2timesjay

Differential Revision: D19956561

fbshipit-source-id: 8d192ecb24858798eaa345ec74229d9c7f0341e1