1/n save_new_trial function
Summary:
Currently, our storage API only exposes one giant "save_experiment" function. This does one of two things:
- If the experiment is new, it creates and saves it
- If the experiment is existing, it performs an equality check over the entire experiment to determine what has changed, and then updates the necessary children.

The logic for the second case is really complicated and expensive. However, we need this behavior because when users are playing with experiments in notebooks, we have no idea what they've changed.

But, in certain cases -- like when calling `get_next_trial` or `complete_trial` in the service API -- we actually do know exactly what has changed, so we can expose more limited update APIs.

Here I'm adding an API that just adds a new trial to an experiment, which we'll be able to use in `get_next_trial` and replace the call here: https://fburl.com/diffusion/lisi8bi5

Reviewed By: stevemandala

Differential Revision: D20260956

fbshipit-source-id: b92cf1eb556442721a7ae425b73bf53827c23ca4