Add cast "transform"
Summary:
See T63950675 for full context. Basically, on RangeParameters, we have a digits parameter which specifies how many digits to round float values to (e.g. 12.34567 --> 12.35). However, we don't apply this rounding to candidates generated from models (or imported arms, for that matter). Hence, when we generate candidates and attach arms, we completely ignore this digits parameter.

To reintroduce this back into Ax (we did this properly in AE), this diff adds
some casting that runs across all parameter values, applied on EVERY
modelbridge while untransforming candidates. We leverage the existing "cast()" function on parameters.

Reviewed By: bletham

Differential Revision: D20441508

fbshipit-source-id: fad03b40ad0a5dd73df515b04f7f042af9b9a9e6