Add more descriptive validation error
Summary:
Currently, if a user tries fitting an MTGP on an experiment with only a single
trial, we do return an error. However, it's a "ChoiceParameter" error which can
be a bit cryptic (quoting past Steve: "Why am I getting a ChoiceParameter error?
I only have RangeParameters?") To make this slight more clear, added a separate
check that has a more relevant error message

Reviewed By: sdsingh

Differential Revision: D20793332

fbshipit-source-id: 9d5cf492940e4e76a5e3d84370fde9d8fbf6306c