Don't enforce max parallelism if `enforce_sequential_optimization=False`
Summary:
This diff does two things: allows for an easy max parallelism override via `choose_generation_strategy` and disables max parallelism enforcement when `enforce_sequential_optimization=False` is passed as a kwarg to `AxClient` (since the whole reason for that kwarg is to "keep generating to matter what", so to keep getting trials even when data is required to proceed to the next model or when a parallelism threshold is reached).

Next step: demonstrate this functionality in the tutorial (T61534200).

Reviewed By: 2timesjay

Differential Revision: D21024015

fbshipit-source-id: 53465780913077f874f35a758c5655c6eb979907