Clean up observation.py
Summary:
I set out to see if there are some easy ways of speeding up `observations_from_data`, but unfortunately that's not the case. Either way, I did some code cleanup in the process, which we may as well land.

Unfortunately, vectorizing this doesn't seem particularly straightforward either, since `ObservationData` contains multiple metrics. The reason this is done is that this data structure will eventually allow to support inputs where we have non-diagonal observation noise covariances. Since we're not really using this capability anywhere at this point, one thing we could potentially consider if this really turns out to be a major bottleneck would be to allow a different kind of of data structure, but that would require a ton of changes throughout and seems like a bad solution.

A better way to deal with this would be to keep a cache of `ObservationData` objects around during the course of the optimization. This would allow to just append new  `ObservationData`  from additional trials to this cache and we'd avoid replicating a ton of effort when getting calling `fetch_data`. We could rebuild that cache when needed (e.g. after a re-loading an experiment). However, it's not entirely clear to me where this cache would live. Presumably it could be on the model and be updated when using the `update` method?

cc ldworkin, bletham, lena-kashtelyan

Reviewed By: lena-kashtelyan

Differential Revision: D21170849

fbshipit-source-id: e5c9970b30680ab7341ec1816b3451cc464fbef9