Add `ttl_hours` property to trials
Summary:
Adding a concept of a "TTL" to a trial, only used if set to a non-null value. In service-like settings it's helpful to know that we shouldn't expect a certain running trial to actually complete because the process that ran it crashed, etc.

This helps with:
1) enforcing parallelism more accurately (else such 'dead' evaluations crowd the available parallelism),
2) being able to not consider dead trials pending (and therefore be able to resuggest the points in them),
3) generally being able to provide full 'scheduler' functionality.

Putting this up to make sure everyone is okay with this before the next steps:
1) Support storage of the TTL in WWW and AOSC,
2) Actually marking the post-TTL trials as failed (currently deciding on the best place to put that logic),
3) Exposing the ability to set TTL in the Service API and scheduler.

Reviewed By: Balandat

Differential Revision: D21095939

fbshipit-source-id: d0eddba9482a38a17abe21c0a203435bea574685