Add modelbridge method for returning transformed points
Summary:
The ModelBridge applies transforms to/from the model space on calls to predict/gen/etc., but there's currently no way to get access to transformed values.

This adds a method to ModelBridge (and Array and Torch subclasses) that allows you to give a set of observation features, and it will return their transformed values in the model space.

The primary purpose for this is to be able to pass in an arbitrary transformed value as a kwarg option to `gen`, for T66200244.

I think this also adds some flexibility for interacting with the model in ad-hoc ways as well. For instance, if I wanted to evaluate the acquisition function on some set of points, there isn't a clear way to do that currently. With this diff, I could take the points, transform them with this method to get the tensor, and then just directly evaluate using `m.model`.

Reviewed By: Balandat

Differential Revision: D21289809

fbshipit-source-id: 8a55bccabe1b4b5923263fdf9ea5ec89ac14504a