Keep `_trial_indices_by_status` mapping on experiment + ensure it stays up to date
Summary:
With the addition of the TTL logic to trials (D21138324), we will need to frequently check the currently running trials and whether they may be past their TTL (for experiments that use TTL). To avoid recomputing the trials by status dict often and introducing a performance sink, I'm adding a mapping to experiment, in which we can keep track of trials by status.

One difficulty this introduces is that the mapping must be kept up-to-date. The logical place to add the update logic is on setting trial status: whenever it changes, the mapping on the experiment should be updated too.

Currently, we discourage setting status directly (our setter: https://fburl.com/diffusion/tp0jwy1u just refuses to set status and tells folks to use `mark_X` methods). I didn't want to remove that exception because folks should use the `mark_X` to change trial statuses, since those methods have some additional important bookkeeping logic.

Instead, I opted for introducing a `_status.setter`, which also requires a `_status` property and the renaming of the `_status` attribute to `__status`. While I don't love the dunder, that seemed like the clearest way to both indicate that the dunder attribute is not to be set manually,

We could also just put the mapping-update logic into the `mark_X` methods, but to me that feels insufficiently fool-proof and like it would be too easy for the mapping to get out of date that way.

Open to other thoughts and alternatives!

Reviewed By: ldworkin

Differential Revision: D21305195

fbshipit-source-id: 578e9871111caf66c15988b753e1e5c0b700ffa8