Mark past-TTL trials as failed
Summary:
Trials, the TTL for which elapsed, need to be marked as `FAILED` when it elapses. It seems to me that the best way to check running trial's TTL is in the experiment's `trials` and `trials_by_status` properties, since those properties will inevitably be called often (and specifically wherever users / higher-level abstractions, like generation strategy, are about to use the information on trial status for anything).

Curious what other folks think on where the right place is to mark failed the past-TTL trials!

Reviewed By: Balandat

Differential Revision: D21138324

fbshipit-source-id: 1d1b73d61f20bcfd3fe4450245af4038a71ede5f