Smarter objective initialization (#319)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/319

Currently Ax will maximizing a metric that had `lower_is_better=True` without any complaints/warnings. I fell victim to this gotcha more than once.

This change now ensures that we emit a warning in this situation.

It also modifies things the constructor to choose appropriate default behavior when passing in a metric with the `lower_is_better` property set.

The behavior if neither a `lower_is_better` property is set on the metric nor a `minimize` arg is provided to the objective remains the same for backward compatibility, but this usage is discouraged via a `DeprecationWarning`. Eventually, I would like to disallow this non-explicit behavior altogether.

Reviewed By: ldworkin

Differential Revision: D21487528

fbshipit-source-id: 9802ef7e939f300bc44cfb5d3ed9fa76e5a9ab22