Instantiate with completed trials' data only if `use_update=True`
Summary:
If the new step is using `update`, it's important to instantiate the new generation step's model with data for completed trials only, so later we can update it with data for new trials as they become completed (since we use the time when trials become `COMPLETED` to identify their data as 'new' data, with which to update the model).

`experiment.fetch_data` will fetch all available data, including for non-completed trials, so to avoid that, we fetch data for specifically completed trials only in `generation_strategy._set_current_model` is `use_update=True` for the step.

Reviewed By: bletham

Differential Revision: D21644324

fbshipit-source-id: ba0e5e476f187ef9bcaf305e379cf095b05b7630