Better error messaging for no exposures on first day of experiment
Summary:
We've known for awhile now that the first PTS report usually fails with a "no exposures" error. This is due to a race condition with the QE job that populates the exposures table. See more context here: https://www.internalfb.com/intern/tasks/?t=58009386&q=1981914255161484&activity=1011457472565422

In this diff, I'm checking if the trial is less than a day old, and if so, throwing a different/ more informative error instead.

Reviewed By: sdsingh

Differential Revision: D21662875

fbshipit-source-id: 39fbef3a2db53ff71586b6ba5cba54d5b18d9bfc