Fix get_logger to avoid double-writing, add formatting of output logfile (#339)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/339

Fixes logging issues reported in T69113923, T69011274.

Essentially, we were adding multiple file handlers to the same loggers.
With this change, the scheduler creates a logger name that includes the memory location, so that it is guaranteed to be unique between multiple instances of the same scheduler class. This prevents cross-contamination of logs. To avoid excessive log outputs, this adds a new `print_name` argument to
`get_logger` that will override the logger name in the logs if provided.

This also adds proper formatting of the output logs.

Reviewed By: stevemandala

Differential Revision: D22271951

fbshipit-source-id: 8a4de9a1b0bd55f22a6bb1348d83bba066e372d0