Retrieve Botorch Acquisition Objective from Ax Logic
Summary:
In the new modular `BoTorchModel` setup, we currently use `get_botorch_objective` (https://fburl.com/diffusion/knqa02pz) to instantiate a BoTorch objective from Ax inputs, here: https://fburl.com/diffusion/tckag10k.

It appears that currently `get_botorch_objective` is incomplete; it doesn't cover the types of objectives needed for NEI (like in this code block: https://fburl.com/diffusion/fr0ndomw).

In this diff, we add a kwarg to this function, specifying whether we should create a `ScalarizedObjective`, rather than a `LinearMCObjective` or a `ConstrainedMCObjective`. We aim to extend `get_botorch_objective` to be more a general-purpose utility that is able to construct an appropriate BoTorch objective type based on Ax inputs.

----

We had a few questions:
- When is `ScalarizedObjective` appropriate?
- Does this kwarg approach seem reasonable?
- Are there any other objective types we should extend `get_botorch_objective` with?

Reviewed By: lena-kashtelyan, qingfeng10

Differential Revision: D22285046

fbshipit-source-id: 62001aa0e8cadfd4561f0c08cf9339da1a8e31cf