MaxValueEntropySearch `Acquisition` with Multi-fidelity Counterpart
Summary:
We want to enroll `MaxValueEntropySearch` and `MultiFidelityMaxValueEntropySearch` into the new modular BoTorch setup. Compared to the `KG` `Acquisition` class, we have a modified `optimize()` function and also retrieve the kwargs `candidate_set` and `maximize`, which `KG` do not need.

-----

NOTE: Using the code in this diff, I tested the setup with N299117. When using `MultiFidelityMaxValueEntropySearch` with `FixedNoiseMultiFidelityGP`, we run into an error. Any suggestions?
```
cholesky_cpu: For batch 2: U(1,1) is zero, singular U.
```

Reviewed By: lena-kashtelyan

Differential Revision: D22459497

fbshipit-source-id: d234cad9b17f754477f627f08b7bd00b6cda2b13