New JSON encoder registry for class Types
Summary:
These changes provide a more automatic way of encoding class `Type`-s.

**Previously:** We had special cases in the encoding logic for `Transform` since we needed to encode `Type[Transform]` and not a `Transform` instance.
**Now:** We generalize the single special `Type[Transform]` case into a case that works for all `Type[...]` encodings.

**Purpose:**
This is necessary as we begin working on JSON storage for the modular `BoTorchModel` and `Surrogate` objects -- here, we need to store `Type[MarginalLogLikelihood]` among a few other `Type`-s.

**Method:**
We use a special `CLASS_ENCODER_REGISTRY` to keep track of all of the `Type[...]` encodings. When encountering a class in `object_to_json`, we iterate through `CLASS_ENCODER_REGISTRY` to call the matching `..._to_dict` function.

What do we think about this approach?

Reviewed By: ldworkin

Differential Revision: D22651305

fbshipit-source-id: 5bf437b4ea99f6c02c024fe25180b99e2f21f923