Re-save experiment to make sure its up-to-date (will update instead in the future)
Summary:
Once experiment is in scheduler, re-save it, since changes might've been made to experiment between the last time it was saved and when it was passed to scheduler (for instance, by forgetting to save experiment after adding metrics, which was happening in our demo notebook for FRL folks).

I think that in the future, it would make sense to just selectively update some fields, but I think we should start doing this ASAP and worry about selective update later.

Reviewed By: ldworkin

Differential Revision: D22821128

fbshipit-source-id: 7530847d2b51e9c388458af86b06c0ede4afee8a