Support building from versioned sdist.
This switches the custom version writer to use `setuptools_scm` which is
a fairly popular method of determining version numbers from git tags
automatically.

Prior to this commit, building an sdist and then building a wheel from
that will use a date-based version, rather than the version of the
sdist.

As an example of the format, here is the version it would have for this
commit before I made it:

```
Version: 0.1.15.dev44+ge1cb797.d20200814
```

and after (it won't have this has anymore, after editing the message,
but you get the idea):

```
Version: 0.1.15.dev45+g5ae1e1c
```