Support building from versioned sdist. (#359)
Summary:
This switches the custom version writer to use `setuptools_scm` which is
a fairly popular method of determining version numbers from git tags
automatically.

Prior to this commit, building an sdist and then building a wheel from
that will use a date-based version, rather than the version of the
sdist.

As an example of the format, here is the version it would have for this
commit before I made it:

```
Version: 0.1.15.dev44+ge1cb797.d20200814
```

and after (it won't have this has anymore, after editing the message,
but you get the idea):

```
Version: 0.1.15.dev45+g5ae1e1c
```

I've done cursory testing that this produces a compatible version string on tags, and what I expect on non-tags; I don't know if you have any CI that relies on a particular format.

Pull Request resolved: https://github.com/facebook/Ax/pull/359

Reviewed By: lena-kashtelyan

Differential Revision: D23161741

Pulled By: ldworkin

fbshipit-source-id: 5533b98d936d7c7b9ce642e1b97d6db9d88caa71