Compute search space param bounds in UnitX
Summary:
Previously, search space bounds were always mapped to [0,1]. This behavior is ok if a model is only used with one search space, the original fitted space. It does not work if the model is used with a new space (e.g. `gen` on a new space, after fitting).

Now, we compute the bounds based on the original search space.

Reviewed By: bletham

Differential Revision: D23114439

fbshipit-source-id: 7a28e941750e1164b84b74c56369d2210a86b63e