If data was logged for a previously unknown metric, add it as tracking
Summary: Something I overlooked before: when data is logged for metrics that aren't on the experiment, we should add those metrics as tracking metrics. This came up through AutoML, where we have been logging data for run costs etc., which caused the experiment to contain data for metrics that weren't actually part of the experiment. This fixes.

Reviewed By: ldworkin

Differential Revision: D23200132

fbshipit-source-id: 7f8271c6b1b941a8994099bbf7c345eb4f89ee33